perf(redis): use Pipelines
add Pipeline support for the publisher: this will increase performance
because publishing will require way less roundtrips to the Redis server.

change up the implementation of the publisher because the pipeline needs
to be flushed every so often. To test out how the pipeline api works, i
decided to go back to a single Producer goroutine that will spawn
another goroutine to flush the pipeline every N messages, where N is a
constant in the publisher function. I'll experiment some more to see if
more publishing goroutines will increase the Pro/Con performance is any
significant way.